// frontend/src/generated/model/usersOutputBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * apportal API
 * Generate application portal — applications, reviews, and the hiring pipeline.
 * OpenAPI spec version: 0.1.0
 */
import type { UsersOutputBodyUsers } from './usersOutputBodyUsers';

export interface UsersOutputBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** True if more results exist beyond this page */
  has_more: boolean;
  users: UsersOutputBodyUsers;
}